RAG/develop/requirements.txt
streamlit>=1.29.0,<2.0.0
openai>=1.0.0,<2.0.0
torch>=2.6.0
sentence-transformers>=3.0.0
transformers>=4.45.0
tokenizers>=0.20.0
chromadb>=0.4.15,<0.5.0
ragas==0.3.7
datasets>=4.0.0
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.23,<0.2.0
langchain-community>=0.0.19,<0.1.0
langchain-openai>=0.0.5,<0.1.0
langchain-text-splitters>=0.0.1,<0.1.0
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0
openpyxl>=3.1.0
pymupdf>=1.23.0
PyPDF2>=3.0.0
Pillow>=10.1.0
python-dotenv>=1.0.0
tqdm>=4.65.0
httpx>=0.24.0,<0.25.0
plotly>=5.17.0
rank-bm25>=0.2.2
appdirs>=1.4.4
diskcache>=5.6.3
instructor>=1.11.0
scikit-network>=0.33.0
nest-asyncio>=1.5.0
huggingface-hub>=0.20.0
